refactor(renderer): move debugging-panel to app/debugging-panel/
Step 4 of the renderer reorganization. The conductor/DHT debugging panel
is dev tooling for the Moss app shell — moves to app/debugging-panel/.

Pure file moves + import-path updates; no behaviour change.
