+++ src/renderer/src/app/debugging-panel/app-debugging-details.ts
@@ -7,8 +7,8 @@ import '@shoelace-style/shoelace/dist/components/card/card.js';
 
 import '../../groups/elements/group-context.js';
 import '../../applets/elements/applet-logo.js';
-import '../../app/dialogs/create-group-dialog.js';
-import '../reusable/groups-for-applet.js';
+import '../dialogs/create-group-dialog.js';
+import '../../elements/reusable/groups-for-applet.js';
 import './cell-details.js';
 
 import { mossStyles } from '../../shared-styles.js';
