+++ src/renderer/src/app/debugging-panel/cell-details.ts
