+++ src/renderer/src/app/debugging-panel/dht-op-detail.ts
