+++ src/renderer/src/app/debugging-panel/memory-chart.ts
